requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
kubernetes>=31.0.0
httpx>=0.27.0
pydantic>=2.9.0
rich>=13.9.0
python-dotenv>=1.0.1
openai>=1.54.0
requests>=2.32.0